Refactor various Ansible playbooks and Terraform configurations for improved clarity and functionality
Refactor removal logic in remover_v2.sh to ensure .terraform directory is only deleted on successful execution

Refactor various scripts for consistency and improved readability, including updates to variable checks and removal of redundant code.

Fix variable reference in bom_download.yaml and improve return value handling in removal scripts

Update deploy/terraform/run/sap_deployer/tfvar_variables.tf


Fix capitalization of "GitHub" in various scripts and update variable descriptions for consistency

Remove duplicate SPN secret description in workload_zone_missing function

+++ deploy/scripts/remover_v2.sh
@@ -764,9 +764,13 @@ function sdaf_remover() {
 		fi
 	fi
 
-	if [ -d ".terraform" ]; then
-		rm -rf ".terraform"
+    if [ "$return_value" -eq 0 ]; then
+
+		if [ -d ".terraform" ]; then
+			rm -rf ".terraform"
+		fi
 	fi
+
 	unset TF_DATA_DIR
 	print_banner "$banner_title" "Removal completed." "info" "Exiting $SCRIPT_NAME"
